Changes README.md, delete src/main.py, delete tests/test_main.py, changed src/__init__.py, changed html-cov, created src/category.py, created src/exceptions.py, created /main.py, created src/product.py, created tests/test_category.py, created tests/test_product.py, created tests/test_exceptions.py.
