+++ htmlcov/index.html
@@ -11,7 +11,7 @@
 <header>
     <div class="content">
         <h1>Coverage report:
-            <span class="pc_cov">88%</span>
+            <span class="pc_cov">92%</span>
         </h1>
         <aside id="help_panel_wrapper">
             <input id="help_panel_state" type="checkbox">
@@ -54,7 +54,7 @@
         </h2>
         <p class="text">
             <a class="nav" href="https://coverage.readthedocs.io/en/7.10.2">coverage.py v7.10.2</a>,
-            created at 2025-09-30 22:52 +1000
+            created at 2025-10-01 19:30 +1000
         </p>
     </div>
 </header>
@@ -72,33 +72,47 @@
         <tbody>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6___init___py.html">src\__init__.py</a></td>
+                <td>4</td>
                 <td>0</td>
                 <td>0</td>
-                <td>0</td>
-                <td class="right" data-ratio="0 0">100%</td>
+                <td class="right" data-ratio="4 4">100%</td>
             </tr>
             <tr class="region">
-                <td class="name left"><a href="z_145eef247bfb46b6_main_py.html">src\main.py</a></td>
-                <td>178</td>
-                <td>24</td>
-                <td>16</td>
-                <td class="right" data-ratio="154 178">87%</td>
+                <td class="name left"><a href="z_145eef247bfb46b6_category_py.html">src\category.py</a></td>
+                <td>82</td>
+                <td>4</td>
+                <td>0</td>
+                <td class="right" data-ratio="78 82">95%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_exceptions_py.html">src\exceptions.py</a></td>
+                <td>4</td>
+                <td>0</td>
+                <td>0</td>
+                <td class="right" data-ratio="4 4">100%</td>
+            </tr>
+            <tr class="region">
+                <td class="name left"><a href="z_145eef247bfb46b6_product_py.html">src\product.py</a></td>
+                <td>95</td>
+                <td>12</td>
+                <td>0</td>
+                <td class="right" data-ratio="83 95">87%</td>
             </tr>
             <tr class="region">
                 <td class="name left"><a href="z_145eef247bfb46b6_utils_py.html">src\utils.py</a></td>
-                <td>20</td>
+                <td>21</td>
                 <td>0</td>
                 <td>10</td>
-                <td class="right" data-ratio="20 20">100%</td>
+                <td class="right" data-ratio="21 21">100%</td>
             </tr>
         </tbody>
         <tfoot>
             <tr class="total">
                 <td class="name left">Total</td>
-                <td>198</td>
-                <td>24</td>
-                <td>26</td>
-                <td class="right" data-ratio="174 198">88%</td>
+                <td>206</td>
+                <td>16</td>
+                <td>10</td>
+                <td class="right" data-ratio="190 206">92%</td>
             </tr>
         </tfoot>
     </table>
@@ -110,7 +124,7 @@
     <div class="content">
         <p>
             <a class="nav" href="https://coverage.readthedocs.io/en/7.10.2">coverage.py v7.10.2</a>,
-            created at 2025-09-30 22:52 +1000
+            created at 2025-10-01 19:30 +1000
         </p>
     </div>
     <aside class="hidden">
