+++ htmlcov/status.json
@@ -1,1 +1,1 @@
-{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.10.2","globals":"a34be2db4021e96e77b3987124bbff70","files":{"z_145eef247bfb46b6___init___py":{"hash":"e6baa73cda2916dad605215f937a92e1","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":0,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_main_py":{"hash":"a2221170e0f310306b54824c853c7380","index":{"url":"z_145eef247bfb46b6_main_py.html","file":"src\\main.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":178,"n_excluded":16,"n_missing":24,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"8580aa1bf10d3a87423fca9dc7bb2246","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":20,"n_excluded":10,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
+{"note":"This file is an internal implementation detail to speed up HTML report generation. Its format can change at any time. You might be looking for the JSON report: https://coverage.rtfd.io/cmd.html#cmd-json","format":5,"version":"7.10.2","globals":"a34be2db4021e96e77b3987124bbff70","files":{"z_145eef247bfb46b6___init___py":{"hash":"25c2228516c7024d4ef2e36b3d4d35fd","index":{"url":"z_145eef247bfb46b6___init___py.html","file":"src\\__init__.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":4,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_main_py":{"hash":"a2221170e0f310306b54824c853c7380","index":{"url":"z_145eef247bfb46b6_main_py.html","file":"src\\main.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":178,"n_excluded":16,"n_missing":24,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_utils_py":{"hash":"35ca796dfef04023cdf8523f7644dd34","index":{"url":"z_145eef247bfb46b6_utils_py.html","file":"src\\utils.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":21,"n_excluded":10,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_category_py":{"hash":"85c5505d84f02126eb5afacbea72d0b0","index":{"url":"z_145eef247bfb46b6_category_py.html","file":"src\\category.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":82,"n_excluded":0,"n_missing":4,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_exceptions_py":{"hash":"5869a75efe75688cef17fe5ee8f39d9e","index":{"url":"z_145eef247bfb46b6_exceptions_py.html","file":"src\\exceptions.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":4,"n_excluded":0,"n_missing":0,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}},"z_145eef247bfb46b6_product_py":{"hash":"5a96686f4b3537eb303d4cc2b7c8da26","index":{"url":"z_145eef247bfb46b6_product_py.html","file":"src\\product.py","description":"","nums":{"precision":0,"n_files":1,"n_statements":95,"n_excluded":0,"n_missing":12,"n_branches":0,"n_partial_branches":0,"n_missing_branches":0}}}}}
